Added correct eventTimes
Almost all times added. Also changed all time variables to double from
int

+++ src/DVD.java
@@ -1,11 +1,11 @@
 
 public class DVD {
 
-	public int startTime = 0;
-	public int endTime = 0;
+	public double startTime = 0;
+	public double endTime = 0;
 	
     //constructor
-    public DVD(int a, int b) {
+    public DVD(double a, double b) {
     	startTime = a;
     	endTime = b;
     }
